+++ gupb/controller/krowa123/pathfinding/__init__.py
@@ -1,3 +1,4 @@
-from .algorithms import astar_search
+from .astar import astar_search
+from .dijkstra import determine_path_bot
 
-__all__ = ["astar_search"]
+__all__ = ["astar_search", "determine_path_bot"]
